ZapRequest moved to its own class ZapReceipt moved to its own class ClassifiedListingTag moved back to ClassifiedListing (as it's technically not a tag)
