+++ nostr-java-api/src/main/java/nostr/api/NIP99.java
@@ -4,7 +4,7 @@ import lombok.NonNull;
 import nostr.api.factory.impl.NIP99Impl.ClassifiedListingEventFactory;
 import nostr.event.BaseTag;
 import nostr.event.NIP99Event;
-import nostr.event.tag.ClassifiedListingTag;
+import nostr.event.tag.ClassifiedListing;
 import nostr.id.Identity;
 
 import java.util.List;
@@ -14,8 +14,8 @@ public class NIP99<T extends NIP99Event> extends EventNostr<T> {
     setSender(sender);
   }
 
-  public NIP99<T> createClassifiedListingEvent(@NonNull List<BaseTag> baseTags, @NonNull String content, @NonNull ClassifiedListingTag classifiedListingTag) {
-    var event = new ClassifiedListingEventFactory(getSender(), baseTags, content, classifiedListingTag).create();
+  public NIP99<T> createClassifiedListingEvent(@NonNull List<BaseTag> baseTags, @NonNull String content, @NonNull ClassifiedListing classifiedListing) {
+    var event = new ClassifiedListingEventFactory(getSender(), baseTags, content, classifiedListing).create();
     setEvent((T) event);
     return this;
   }
