+++ nostr-java-test/src/test/java/nostr/test/json/JsonParseTest.java
@@ -11,7 +11,6 @@ import nostr.event.BaseEvent;
 import nostr.event.BaseMessage;
 import nostr.event.BaseTag;
 import nostr.event.Marker;
-import nostr.event.impl.ClassifiedListingEvent;
 import nostr.event.impl.Filters;
 import nostr.event.impl.GenericEvent;
 import nostr.event.impl.GenericTag;
@@ -22,7 +21,6 @@ import nostr.event.list.KindList;
 import nostr.event.list.PublicKeyList;
 import nostr.event.message.EventMessage;
 import nostr.event.message.ReqMessage;
-import nostr.event.tag.ClassifiedListingTag;
 import nostr.event.tag.EventTag;
 import nostr.event.tag.PriceTag;
 import nostr.event.tag.PubKeyTag;
